refactor(AresCachedManager): rename file to match class name
