+++ src/lib/classes/cachedManager.ts
