+++ src/lib/classes/index.ts
@@ -1,7 +1,7 @@
 export * from "./base";
 export * from "./baseManager";
 export * from "./baseModule";
-export * from "./cacheManager";
+export * from "./cachedManager";
 export * from "./cacheResults";
 export * from "./client";
 export * from "./error";
